further security enhancements

+++ front-service/src/main/java/com/isc/project/manager/service/ProjectProxyService.java
@@ -2,9 +2,9 @@ package com.isc.project.manager.service;
 
 import com.isc.project.manager.api.dto.ProjectDTO;
 import com.isc.project.manager.persistence.domain.UserType;
-import com.isc.project.manager.security.authorization.AuthenticatedUserHolder;
+import com.isc.project.manager.security.authorization.AuthorizationService;
 import com.isc.project.manager.security.authorization.SecurityActionType;
-import com.isc.project.manager.security.authorization.SecurityDecoratorService;
+import com.isc.project.manager.security.authorization.UserHolder;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.http.HttpEntity;
@@ -20,15 +20,15 @@ import java.util.stream.Collectors;
 public class ProjectProxyService {
     private final RestTemplate restTemplate = new RestTemplate();
     private final String serverAddress;
-    private final SecurityDecoratorService securityDecoratorService;
-    private final AuthenticatedUserHolder userHolder;
+    private final AuthorizationService authorizationService;
+    private final UserHolder userHolder;
 
     @Autowired
     public ProjectProxyService(@Qualifier("projectApiServiceAddress") String serverAddress,
-                               SecurityDecoratorService securityDecoratorService,
-                               AuthenticatedUserHolder userHolder) {
+                               AuthorizationService authorizationService,
+                               UserHolder userHolder) {
         this.serverAddress = serverAddress;
-        this.securityDecoratorService = securityDecoratorService;
+        this.authorizationService = authorizationService;
         this.userHolder = userHolder;
     }
 
@@ -39,26 +39,29 @@ public class ProjectProxyService {
                 : userHolder.getCurrentUser().getTenantCode();
 
         return findAllByTenant(tenantCode).stream()
-                .map(securityDecoratorService::executeReadPolicy)
+                .filter(projectDTO -> authorizationService.canAccess(projectDTO, SecurityActionType.READ))
                 .collect(Collectors.toList());
     }
 
     private List<ProjectDTO> findAllByTenant(String tenantCode) {
-        return Arrays.asList(restTemplate.getForObject(
+        return Arrays.stream(restTemplate.getForObject(
                 serverAddress + "/api/projects" + (tenantCode.isEmpty() ? "" : "?tenant=" + tenantCode),
-                ProjectDTO[].class
-        ));
+                ProjectDTO[].class))
+                .filter(projectDTO -> authorizationService.canAccess(projectDTO, SecurityActionType.READ))
+                .collect(Collectors.toList());
     }
 
     public ProjectDTO getById(Long id) {
         ProjectDTO projectDTO = restTemplate.getForObject(serverAddress + "/api/projects/" + id, ProjectDTO.class);
+        authorizationService.checkAccess(projectDTO, SecurityActionType.READ);
 
-        return securityDecoratorService.executeReadPolicy(projectDTO);
+        return projectDTO;
     }
 
     public ProjectDTO create(ProjectDTO projectDTO) {
         projectDTO.setId(null);
-        securityDecoratorService.executeCreatePolicy(projectDTO);
+        projectDTO.setTenantCode(userHolder.getCurrentTenant());
+        authorizationService.checkAccess(projectDTO, SecurityActionType.CREATE);
 
         return restTemplate.postForObject(
                 serverAddress + "/api/projects",
@@ -66,12 +69,16 @@ public class ProjectProxyService {
                 ProjectDTO.class);
     }
 
-    public ProjectDTO update(Long id, ProjectDTO projectDTO) {
-        securityDecoratorService.executeUpdatePolicy(getById(id));
+    public ProjectDTO update(ProjectDTO newDTO) {
+        ProjectDTO oldDTO = getById(newDTO.getId());
+        authorizationService.checkAccess(oldDTO, SecurityActionType.UPDATE);
 
-        HttpEntity<ProjectDTO> requestEntity = new HttpEntity<>(projectDTO);
+        newDTO.setId(oldDTO.getId());
+        newDTO.setTenantCode(oldDTO.getTenantCode());
+
+        HttpEntity<ProjectDTO> requestEntity = new HttpEntity<>(newDTO);
         return restTemplate.exchange(
-                serverAddress + "/api/projects/" + id,
+                serverAddress + "/api/projects/" + newDTO.getId(),
                 HttpMethod.PUT,
                 requestEntity,
                 ProjectDTO.class
@@ -79,7 +86,7 @@ public class ProjectProxyService {
     }
 
     public void delete(Long id) {
-        securityDecoratorService.executeDeletePolicy(getById(id));
+        authorizationService.checkAccess(getById(id), SecurityActionType.DELETE);
 
         restTemplate.delete(serverAddress + "/api/projects/" + id);
     }
